Inline remaining replay.go function and remove stale .yaml replay files
