+++ internal/testing/setup/setup.go
@@ -33,6 +33,7 @@ import (
 	"gocloud.dev/internal/useragent"
 
 	"cloud.google.com/go/httpreplay"
+	"cloud.google.com/go/rpcreplay"
 	"golang.org/x/oauth2/google"
 	"google.golang.org/api/option"
 	"google.golang.org/grpc"
@@ -296,4 +297,41 @@ func FakeGCPDefaultCredentials(t *testing.T) func() {
 		os.Remove(f.Name())
 		os.Setenv(envVar, oldEnvVal)
 	}
+}
+
+// NewGCPDialOptions return grpc.DialOptions that are to be appended to a GRPC
+// dial request. These options allow a recorder/replayer to intercept RPCs and
+// save RPCs to the file at filename, or read the RPCs from the file and return
+// them. When recording is set to true, we're in recording mode; otherwise we're
+// in replaying mode.
+func NewGCPDialOptions(t *testing.T, recording bool, filename string) (opts []grpc.DialOption, done func()) {
+	path := filepath.Join("testdata", filename)
+	if recording {
+		t.Logf("Recording into golden file %s", path)
+		r, err := rpcreplay.NewRecorder(path, nil)
+		if err != nil {
+			t.Fatal(err)
+		}
+		opts = r.DialOptions()
+		done = func() {
+			if err := r.Close(); err != nil {
+				t.Errorf("unable to close recorder: %v", err)
+			}
+		}
+		return opts, done
+	}
+	t.Logf("Replaying from golden file %s", path)
+	r, err := rpcreplay.NewReplayer(path)
+	if err != nil {
+		t.Fatal(err)
+	}
+	// Uncomment for more verbose logging from the replayer.
+	// r.SetLogFunc(t.Logf)
+	opts = r.DialOptions()
+	done = func() {
+		if err := r.Close(); err != nil {
+			t.Errorf("unable to close recorder: %v", err)
+		}
+	}
+	return opts, done
 }
