--- secrets/awskms/testdata/TestConformance/TestAs/verify_As_function.yaml
@@ -1,31 +1,0 @@
----
-version: 1
-interactions:
-- request:
-    body: '{"CiphertextBlob":"bWFsZm9ybWVkIGNpcGhlciBtZXNzYWdl"}'
-    form: {}
-    headers:
-      Content-Length:
-      - "53"
-      Content-Type:
-      - application/x-amz-json-1.1
-      User-Agent:
-      - aws-sdk-go/1.16.23 (go1.11.2; darwin; amd64)
-      X-Amz-Date:
-      - 20190207T222827Z
-      X-Amz-Target:
-      - TrentService.Decrypt
-    url: https://kms.us-east-2.amazonaws.com/
-    method: POST
-  response:
-    body: '{"__type":"InvalidCiphertextException"}'
-    headers:
-      Content-Length:
-      - "39"
-      Content-Type:
-      - application/x-amz-json-1.1
-      X-Amzn-Requestid:
-      - fa0679a4-6d47-4015-a600-2f96e0439d10
-    status: 400 Bad Request
-    code: 400
-    duration: ""
