--- secrets/awskms/testdata/TestConformance/TestAs/verify_As_returns_false_when_passed_nil.yaml
@@ -1,31 +1,0 @@
----
-version: 1
-interactions:
-- request:
-    body: '{"CiphertextBlob":"bWFsZm9ybWVkIGNpcGhlciBtZXNzYWdl"}'
-    form: {}
-    headers:
-      Content-Length:
-      - "53"
-      Content-Type:
-      - application/x-amz-json-1.1
-      User-Agent:
-      - aws-sdk-go/1.16.23 (go1.11.2; darwin; amd64)
-      X-Amz-Date:
-      - 20190208T190450Z
-      X-Amz-Target:
-      - TrentService.Decrypt
-    url: https://kms.us-east-2.amazonaws.com/
-    method: POST
-  response:
-    body: '{"__type":"InvalidCiphertextException"}'
-    headers:
-      Content-Length:
-      - "39"
-      Content-Type:
-      - application/x-amz-json-1.1
-      X-Amzn-Requestid:
-      - 13e6a637-d033-45e3-a1a7-b76808873658
-    status: 400 Bad Request
-    code: 400
-    duration: ""
