--- secrets/awskms/testdata/TestConformance/TestDecryptMalformedError.yaml
@@ -1,31 +1,0 @@
----
-version: 1
-interactions:
-- request:
-    body: '{"CiphertextBlob":"bWFsZm9ybWVkIGNpcGhlciBtZXNzYWdl"}'
-    form: {}
-    headers:
-      Content-Length:
-      - "53"
-      Content-Type:
-      - application/x-amz-json-1.1
-      User-Agent:
-      - aws-sdk-go/1.16.23 (go1.11.2; linux; amd64)
-      X-Amz-Date:
-      - 20190131T192205Z
-      X-Amz-Target:
-      - TrentService.Decrypt
-    url: https://kms.us-east-2.amazonaws.com/
-    method: POST
-  response:
-    body: '{"__type":"InvalidCiphertextException"}'
-    headers:
-      Content-Length:
-      - "39"
-      Content-Type:
-      - application/x-amz-json-1.1
-      X-Amzn-Requestid:
-      - bc7e5b6b-cfce-4da8-86ce-90c371eeb487
-    status: 400 Bad Request
-    code: 400
-    duration: ""
