--- secrets/awskms/testdata/TestConformance/TestEncryptDecrypt.yaml
@@ -1,59 +1,0 @@
----
-version: 1
-interactions:
-- request:
-    body: '{"KeyId":"alias/test-secrets","Plaintext":"SSdtIGEgc2VjcmV0IG1lc3NhZ2Uh"}'
-    form: {}
-    headers:
-      Content-Length:
-      - "73"
-      Content-Type:
-      - application/x-amz-json-1.1
-      User-Agent:
-      - aws-sdk-go/1.16.23 (go1.11.2; linux; amd64)
-      X-Amz-Date:
-      - 20190131T192204Z
-      X-Amz-Target:
-      - TrentService.Encrypt
-    url: https://kms.us-east-2.amazonaws.com/
-    method: POST
-  response:
-    body: '{"CiphertextBlob":"AQICAHgPAmswqcK5X3uDPoAQGriWQltcXwEzOp2BOdwxo57cVwEK/AaPgIxb+HiP2i+XYfJhAAAAczBxBgkqhkiG9w0BBwagZDBiAgEAMF0GCSqGSIb3DQEHATAeBglghkgBZQMEAS4wEQQMUv0W0E0i3Hs1OOHuAgEQgDD4DxT5QlS0bFq1We6bvUO5J40R/6tiZ2OdvKHNxzHFEeCmQQgH3RjJnsNNEpIkGMQ=","KeyId":"arn:aws:kms:us-east-2:462380225722:key/89388a7e-51ac-4fa5-a1e5-042141ef026c"}'
-    headers:
-      Content-Length:
-      - "339"
-      Content-Type:
-      - application/x-amz-json-1.1
-      X-Amzn-Requestid:
-      - 7823d50f-a9d8-48a3-b43a-6c499e23a7a6
-    status: 200 OK
-    code: 200
-    duration: ""
-- request:
-    body: '{"CiphertextBlob":"AQICAHgPAmswqcK5X3uDPoAQGriWQltcXwEzOp2BOdwxo57cVwEK/AaPgIxb+HiP2i+XYfJhAAAAczBxBgkqhkiG9w0BBwagZDBiAgEAMF0GCSqGSIb3DQEHATAeBglghkgBZQMEAS4wEQQMUv0W0E0i3Hs1OOHuAgEQgDD4DxT5QlS0bFq1We6bvUO5J40R/6tiZ2OdvKHNxzHFEeCmQQgH3RjJnsNNEpIkGMQ="}'
-    form: {}
-    headers:
-      Content-Length:
-      - "253"
-      Content-Type:
-      - application/x-amz-json-1.1
-      User-Agent:
-      - aws-sdk-go/1.16.23 (go1.11.2; linux; amd64)
-      X-Amz-Date:
-      - 20190131T192204Z
-      X-Amz-Target:
-      - TrentService.Decrypt
-    url: https://kms.us-east-2.amazonaws.com/
-    method: POST
-  response:
-    body: '{"KeyId":"arn:aws:kms:us-east-2:462380225722:key/89388a7e-51ac-4fa5-a1e5-042141ef026c","Plaintext":"SSdtIGEgc2VjcmV0IG1lc3NhZ2Uh"}'
-    headers:
-      Content-Length:
-      - "130"
-      Content-Type:
-      - application/x-amz-json-1.1
-      X-Amzn-Requestid:
-      - 2928527b-fdf6-48d0-bbf9-50943edd5933
-    status: 200 OK
-    code: 200
-    duration: ""
