--- secrets/awskms/testdata/TestConformance/TestMultipleEncryptionsNotEqual.yaml
@@ -1,115 +1,0 @@
----
-version: 1
-interactions:
-- request:
-    body: '{"KeyId":"alias/test-secrets","Plaintext":"SSdtIGEgc2VjcmV0IG1lc3NhZ2Uh"}'
-    form: {}
-    headers:
-      Content-Length:
-      - "73"
-      Content-Type:
-      - application/x-amz-json-1.1
-      User-Agent:
-      - aws-sdk-go/1.16.23 (go1.11.2; linux; amd64)
-      X-Amz-Date:
-      - 20190131T192204Z
-      X-Amz-Target:
-      - TrentService.Encrypt
-    url: https://kms.us-east-2.amazonaws.com/
-    method: POST
-  response:
-    body: '{"CiphertextBlob":"AQICAHgPAmswqcK5X3uDPoAQGriWQltcXwEzOp2BOdwxo57cVwGc9GGlkYcI7d6c2maQpKMgAAAAczBxBgkqhkiG9w0BBwagZDBiAgEAMF0GCSqGSIb3DQEHATAeBglghkgBZQMEAS4wEQQMTYHcToki5KuA00vnAgEQgDCeH2AgTJmrWDjXyOTWhcphSYDehHWbzPFbTXP6HUPlj9GXd5hqF59RsNJM7UGNqYE=","KeyId":"arn:aws:kms:us-east-2:462380225722:key/89388a7e-51ac-4fa5-a1e5-042141ef026c"}'
-    headers:
-      Content-Length:
-      - "339"
-      Content-Type:
-      - application/x-amz-json-1.1
-      X-Amzn-Requestid:
-      - 8284d7b3-82da-4473-a583-3bdae5386cf7
-    status: 200 OK
-    code: 200
-    duration: ""
-- request:
-    body: '{"KeyId":"alias/test-secrets","Plaintext":"SSdtIGEgc2VjcmV0IG1lc3NhZ2Uh"}'
-    form: {}
-    headers:
-      Content-Length:
-      - "73"
-      Content-Type:
-      - application/x-amz-json-1.1
-      User-Agent:
-      - aws-sdk-go/1.16.23 (go1.11.2; linux; amd64)
-      X-Amz-Date:
-      - 20190131T192204Z
-      X-Amz-Target:
-      - TrentService.Encrypt
-    url: https://kms.us-east-2.amazonaws.com/
-    method: POST
-  response:
-    body: '{"CiphertextBlob":"AQICAHgPAmswqcK5X3uDPoAQGriWQltcXwEzOp2BOdwxo57cVwF+o4kCxqicu9OvIznor8iGAAAAczBxBgkqhkiG9w0BBwagZDBiAgEAMF0GCSqGSIb3DQEHATAeBglghkgBZQMEAS4wEQQMn7fq6uP8Q3rNxsQeAgEQgDAGM2VpL+clkaIg1H3QcJrXKYlZ9hgcJ5fTo0ZfInlTxGDPUG+rnYMMoshJ8RfD8GM=","KeyId":"arn:aws:kms:us-east-2:462380225722:key/89388a7e-51ac-4fa5-a1e5-042141ef026c"}'
-    headers:
-      Content-Length:
-      - "339"
-      Content-Type:
-      - application/x-amz-json-1.1
-      X-Amzn-Requestid:
-      - 081a5078-00af-4de3-81ec-8319117fbc06
-    status: 200 OK
-    code: 200
-    duration: ""
-- request:
-    body: '{"CiphertextBlob":"AQICAHgPAmswqcK5X3uDPoAQGriWQltcXwEzOp2BOdwxo57cVwGc9GGlkYcI7d6c2maQpKMgAAAAczBxBgkqhkiG9w0BBwagZDBiAgEAMF0GCSqGSIb3DQEHATAeBglghkgBZQMEAS4wEQQMTYHcToki5KuA00vnAgEQgDCeH2AgTJmrWDjXyOTWhcphSYDehHWbzPFbTXP6HUPlj9GXd5hqF59RsNJM7UGNqYE="}'
-    form: {}
-    headers:
-      Content-Length:
-      - "253"
-      Content-Type:
-      - application/x-amz-json-1.1
-      User-Agent:
-      - aws-sdk-go/1.16.23 (go1.11.2; linux; amd64)
-      X-Amz-Date:
-      - 20190131T192205Z
-      X-Amz-Target:
-      - TrentService.Decrypt
-    url: https://kms.us-east-2.amazonaws.com/
-    method: POST
-  response:
-    body: '{"KeyId":"arn:aws:kms:us-east-2:462380225722:key/89388a7e-51ac-4fa5-a1e5-042141ef026c","Plaintext":"SSdtIGEgc2VjcmV0IG1lc3NhZ2Uh"}'
-    headers:
-      Content-Length:
-      - "130"
-      Content-Type:
-      - application/x-amz-json-1.1
-      X-Amzn-Requestid:
-      - b94d4194-b533-42ba-8019-5b607cd4a2bc
-    status: 200 OK
-    code: 200
-    duration: ""
-- request:
-    body: '{"CiphertextBlob":"AQICAHgPAmswqcK5X3uDPoAQGriWQltcXwEzOp2BOdwxo57cVwF+o4kCxqicu9OvIznor8iGAAAAczBxBgkqhkiG9w0BBwagZDBiAgEAMF0GCSqGSIb3DQEHATAeBglghkgBZQMEAS4wEQQMn7fq6uP8Q3rNxsQeAgEQgDAGM2VpL+clkaIg1H3QcJrXKYlZ9hgcJ5fTo0ZfInlTxGDPUG+rnYMMoshJ8RfD8GM="}'
-    form: {}
-    headers:
-      Content-Length:
-      - "253"
-      Content-Type:
-      - application/x-amz-json-1.1
-      User-Agent:
-      - aws-sdk-go/1.16.23 (go1.11.2; linux; amd64)
-      X-Amz-Date:
-      - 20190131T192205Z
-      X-Amz-Target:
-      - TrentService.Decrypt
-    url: https://kms.us-east-2.amazonaws.com/
-    method: POST
-  response:
-    body: '{"KeyId":"arn:aws:kms:us-east-2:462380225722:key/89388a7e-51ac-4fa5-a1e5-042141ef026c","Plaintext":"SSdtIGEgc2VjcmV0IG1lc3NhZ2Uh"}'
-    headers:
-      Content-Length:
-      - "130"
-      Content-Type:
-      - application/x-amz-json-1.1
-      X-Amzn-Requestid:
-      - 87e72b7d-368a-4d60-b445-86cade4ec4fa
-    status: 200 OK
-    code: 200
-    duration: ""
