--- secrets/awskms/testdata/TestConformance/TestMultipleKeys.yaml
@@ -1,115 +1,0 @@
----
-version: 1
-interactions:
-- request:
-    body: '{"KeyId":"alias/test-secrets","Plaintext":"SSdtIGEgc2VjcmV0IG1lc3NhZ2Uh"}'
-    form: {}
-    headers:
-      Content-Length:
-      - "73"
-      Content-Type:
-      - application/x-amz-json-1.1
-      User-Agent:
-      - aws-sdk-go/1.16.23 (go1.11.2; linux; amd64)
-      X-Amz-Date:
-      - 20190131T192205Z
-      X-Amz-Target:
-      - TrentService.Encrypt
-    url: https://kms.us-east-2.amazonaws.com/
-    method: POST
-  response:
-    body: '{"CiphertextBlob":"AQICAHgPAmswqcK5X3uDPoAQGriWQltcXwEzOp2BOdwxo57cVwGmXOsWkCWO3/GSB0rvLaWEAAAAczBxBgkqhkiG9w0BBwagZDBiAgEAMF0GCSqGSIb3DQEHATAeBglghkgBZQMEAS4wEQQMmaBGhINtNj/1WQqjAgEQgDBfeiVsZVDww0k6qnyJyX008xIhyOzs5sWkKdeoQgkdS07U8H65oWqxGOCF82rpclM=","KeyId":"arn:aws:kms:us-east-2:462380225722:key/89388a7e-51ac-4fa5-a1e5-042141ef026c"}'
-    headers:
-      Content-Length:
-      - "339"
-      Content-Type:
-      - application/x-amz-json-1.1
-      X-Amzn-Requestid:
-      - 89d560fc-dfa3-415d-bb4e-f9390f89f62e
-    status: 200 OK
-    code: 200
-    duration: ""
-- request:
-    body: '{"KeyId":"alias/test-secrets2","Plaintext":"SSdtIGEgc2VjcmV0IG1lc3NhZ2Uh"}'
-    form: {}
-    headers:
-      Content-Length:
-      - "74"
-      Content-Type:
-      - application/x-amz-json-1.1
-      User-Agent:
-      - aws-sdk-go/1.16.23 (go1.11.2; linux; amd64)
-      X-Amz-Date:
-      - 20190131T192205Z
-      X-Amz-Target:
-      - TrentService.Encrypt
-    url: https://kms.us-east-2.amazonaws.com/
-    method: POST
-  response:
-    body: '{"CiphertextBlob":"AQICAHglssInw5+kwMxa6r95YoczBbK5KxmDVGO+ZZE6InreVwFHCoej6l/wwO/KAxCprDOgAAAAczBxBgkqhkiG9w0BBwagZDBiAgEAMF0GCSqGSIb3DQEHATAeBglghkgBZQMEAS4wEQQMrtCnuyBAwbLrEHBZAgEQgDAwUj5bNi5loSqOY1BhcutmUWiCQBc7lehwu4vMUCYIZ2b717iNQ7fDuJ68WRFXqD8=","KeyId":"arn:aws:kms:us-east-2:462380225722:key/f6c62289-1361-4c78-b7b6-8eae619494d2"}'
-    headers:
-      Content-Length:
-      - "339"
-      Content-Type:
-      - application/x-amz-json-1.1
-      X-Amzn-Requestid:
-      - c29de524-7d33-4c57-9412-95335537ddc3
-    status: 200 OK
-    code: 200
-    duration: ""
-- request:
-    body: '{"CiphertextBlob":"AQICAHgPAmswqcK5X3uDPoAQGriWQltcXwEzOp2BOdwxo57cVwGmXOsWkCWO3/GSB0rvLaWEAAAAczBxBgkqhkiG9w0BBwagZDBiAgEAMF0GCSqGSIb3DQEHATAeBglghkgBZQMEAS4wEQQMmaBGhINtNj/1WQqjAgEQgDBfeiVsZVDww0k6qnyJyX008xIhyOzs5sWkKdeoQgkdS07U8H65oWqxGOCF82rpclM="}'
-    form: {}
-    headers:
-      Content-Length:
-      - "253"
-      Content-Type:
-      - application/x-amz-json-1.1
-      User-Agent:
-      - aws-sdk-go/1.16.23 (go1.11.2; linux; amd64)
-      X-Amz-Date:
-      - 20190131T192205Z
-      X-Amz-Target:
-      - TrentService.Decrypt
-    url: https://kms.us-east-2.amazonaws.com/
-    method: POST
-  response:
-    body: '{"KeyId":"arn:aws:kms:us-east-2:462380225722:key/89388a7e-51ac-4fa5-a1e5-042141ef026c","Plaintext":"SSdtIGEgc2VjcmV0IG1lc3NhZ2Uh"}'
-    headers:
-      Content-Length:
-      - "130"
-      Content-Type:
-      - application/x-amz-json-1.1
-      X-Amzn-Requestid:
-      - 9a3c6e43-4b4f-4ac0-bcd5-ffffd6471419
-    status: 200 OK
-    code: 200
-    duration: ""
-- request:
-    body: '{"CiphertextBlob":"AQICAHglssInw5+kwMxa6r95YoczBbK5KxmDVGO+ZZE6InreVwFHCoej6l/wwO/KAxCprDOgAAAAczBxBgkqhkiG9w0BBwagZDBiAgEAMF0GCSqGSIb3DQEHATAeBglghkgBZQMEAS4wEQQMrtCnuyBAwbLrEHBZAgEQgDAwUj5bNi5loSqOY1BhcutmUWiCQBc7lehwu4vMUCYIZ2b717iNQ7fDuJ68WRFXqD8="}'
-    form: {}
-    headers:
-      Content-Length:
-      - "253"
-      Content-Type:
-      - application/x-amz-json-1.1
-      User-Agent:
-      - aws-sdk-go/1.16.23 (go1.11.2; linux; amd64)
-      X-Amz-Date:
-      - 20190131T192205Z
-      X-Amz-Target:
-      - TrentService.Decrypt
-    url: https://kms.us-east-2.amazonaws.com/
-    method: POST
-  response:
-    body: '{"KeyId":"arn:aws:kms:us-east-2:462380225722:key/f6c62289-1361-4c78-b7b6-8eae619494d2","Plaintext":"SSdtIGEgc2VjcmV0IG1lc3NhZ2Uh"}'
-    headers:
-      Content-Length:
-      - "130"
-      Content-Type:
-      - application/x-amz-json-1.1
-      X-Amzn-Requestid:
-      - 7177ddd6-44a3-4e2c-a7a9-9ea18a69a3b3
-    status: 200 OK
-    code: 200
-    duration: ""
